relocate generator classes and update imports
